ai-song-workshop/requirements.txt
# Core web framework
flask==3.0.0
flask-cors==4.0.0

# API clients
anthropic>=0.40.0
requests==2.31.0
google-genai==1.0.0

# Template engine (included with Flask)
jinja2==3.1.3

# Session management
flask-session==0.6.0

# Environment variables
python-dotenv==1.0.1

# Testing
pytest==7.4.3
pytest-asyncio==0.23.4
hypothesis==6.98.0

# Async support
aiohttp==3.9.3

# Image processing
Pillow>=10.3.0
